+++ packages/brick-kit/src/core/CustomTemplates.spec.ts
@@ -260,9 +260,7 @@ describe("handleProxyOfCustomTemplate", () => {
             ref: "button",
             refTransform: {
               buttonType: "<% DATA.isDanger ? 'danger' : 'default' %>",
-              style: {
-                display: "<% DATA.isDanger ? 'inline' : 'block' %>",
-              },
+              'style.display': "<% DATA.isDanger ? 'inline' : 'block' %>",
             },
           },
         },