+++ packages/brick-kit/src/core/LocationContext.ts
@@ -1,4 +1,4 @@
-import { omit } from "lodash";
+import { omit, set } from "lodash";
 import {
   PluginLocation,
   MatchResult,
@@ -143,6 +143,7 @@ export class LocationContext {
           await this.resolver.resolveOne(
             "reference",
             {
+              transform: "value",
               transformMapArray: false,
               ...contextConf.resolve,
             },
@@ -382,7 +383,7 @@ export class LocationContext {
     const otherMenuConf = omit(menuConf, ["injectDeep", "type"]);
     const injectedMenuConf =
       injectDeep !== false
-        ? computeRealProperties(otherMenuConf, context, true)
+        ? computeRealValue(otherMenuConf, context, true)
         : otherMenuConf;
     const {
       sidebarMenu,
@@ -533,13 +534,10 @@ export class LocationContext {
 
     Object.assign(brick, {
       type: tplTagName || brickConf.brick,
-      properties: Object.assign(
-        computeRealProperties(
-          brickConf.properties,
-          context,
-          brickConf.injectDeep !== false
-        ),
-        (brickConf as RuntimeBrickConf).$$computedPropsFromProxy
+      properties: computeRealProperties(
+        brickConf.properties,
+        context,
+        brickConf.injectDeep !== false
       ),
       events: isObject(brickConf.events) ? brickConf.events : {},
       context,
@@ -548,6 +546,14 @@ export class LocationContext {
       refForProxy: (brickConf as RuntimeBrickConf).$$refForProxy,
       parentTemplate: (brickConf as RuntimeBrickConf).$$parentTemplate,
     });
+
+    if ((brickConf as RuntimeBrickConf).$$computedPropsFromProxy) {
+      Object.entries(
+        (brickConf as RuntimeBrickConf).$$computedPropsFromProxy
+      ).forEach(([propName, propValue]) => {
+        set(brick.properties, propName, propValue);
+      });
+    }
 
     if (brick.refForProxy) {
       brick.refForProxy.brick = brick;